requirements.txt
asgiref==3.7.2
bootstrap4==0.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==7.1.2
Django==4.2.4
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
docker==4.2.0
dockerpty==0.4.1
docopt==0.6.2
gitdb==4.0.10
GitPython==3.1.0
idna==3.4
mypy==1.5.0
mypy-extensions==1.0.0
popper==2.7.0
pykwalify==1.7.0
pypiwin32==223
python-box==4.2.3
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==5.3.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
semver==3.0.1
six==1.16.0
smmap==5.0.0
sorl-thumbnail==12.9.0
spython==0.0.79
sqlparse==0.4.4
thumbnail==1.5
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.6.1